parallelizing marching cubes

+++ slime/simulation/marching_cubes_tables.cpp
@@ -2,5 +2,5 @@
 
 using namespace slime;
 
-constexpr int8_t MarchingCubesTables::cornerIndexFromEdge[12][2];
-constexpr int8_t MarchingCubesTables::triangulation[256][16];
+constexpr int MarchingCubesTables::cornerIndexFromEdge[12][2];
+constexpr int MarchingCubesTables::triangulation[256][16];
